Move CBanDB from net.h|.cpp to bandb.h|.cpp
Include addition of doxygen commments and enforcing clang-format.
